requirements.txt
# Framework Universal Niche V1.5 Bot — Dependencies
# Build by Fingercod3

# AI
anthropic>=0.40.0

# Telegram
python-telegram-bot[job-queue]>=21.0

# Discord
discord.py>=2.3.0

# Session storage
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
